feat: moved useDefaultStyles hook to own file
