+++ app/hooks.ts
@@ -14,45 +14,10 @@ import {
   RIGHT_TEXT
 } from "./constants"
 import useDebounce from "./hooks/useDebounce"
+import useDefaultStyles from "./hooks/useDefaultStyles"
 import useElements from "./hooks/useElements"
 import useFonts from "./hooks/useFonts"
 import type { SelectorProps, StylesData } from "./types"
-
-export const useDefaultStyles = () => {
-  const elements = useElements()
-
-  const styles = useRef({
-    [GOAL_WIDGET]: {
-      "background-color": elements[GOAL_WIDGET].style.backgroundColor,
-      "border-color": elements[GOAL_WIDGET].style.borderColor,
-      "border-width": elements[GOAL_WIDGET].style.borderWidth,
-      "border-radius": elements[GOAL_WIDGET].style.borderRadius || "1rem"
-    },
-    [GW_PROGRESS_BAR]: {
-      "background-color": elements[GW_PROGRESS_BAR].style.backgroundColor
-    },
-    [LEFT_TEXT]: {
-      color: elements[LEFT_TEXT].style.color,
-      "font-size": elements[LEFT_TEXT].style.fontSize || "18px",
-      "font-family": elements[LEFT_TEXT].style.fontFamily || "Inter"
-    },
-    [RIGHT_TEXT]: {
-      color: elements[RIGHT_TEXT].style.color,
-      "font-size": elements[RIGHT_TEXT].style.fontSize || "18px",
-      "font-family": elements[RIGHT_TEXT].style.fontFamily || "Inter"
-    },
-    [GW_IMAGE]: {
-      content: `url(${elements[GW_IMAGE].src})`
-    }
-  } as StylesData)
-
-  const [defaultStyles] = useStorage(
-    "defaultStyles",
-    (value?: StylesData) => value ?? styles.current
-  )
-
-  return defaultStyles
-}
 
 export const useStyles = () => {
   const elements = useElements()
